+++ adapters/swi.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2012 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for SWI Prolog 6.0.0 and later versions
-%  Last updated on October 21, 2012
+%  Last updated on November 28, 2012
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -967,21 +967,6 @@ user:goal_expansion('::'(Object, Message), user:Goal) :-
 %  message token printing
 %
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-
-:- multifile('$lgt_logtalk.message_hook'/5).
-:- dynamic('$lgt_logtalk.message_hook'/5).
-
-'$lgt_logtalk.message_hook'(_, Kind, Component, Tokens, ExCtx) :-
-	% make sure we're no running in PDT as this IDE also defines
-	% this hook predicate for much improved Logtalk integration
-	\+ current_module(pdt_console),
-	% Logtalk message kinds can be either an atom or a compound
-	% term but SWI-Prolog only uses atoms
-	functor(Kind, Functor, _),
-	'$lgt_append'([begin(Functor,Ctx)| Tokens], [end(Ctx)], ExpandedTokens),
-	'$lgt_logtalk.message_prefix_stream'(Kind, Component, Prefix, Stream, ExCtx),
-	'$lgt_logtalk.print_message_tokens'(Stream, Prefix, ExpandedTokens, ExCtx).
 
 
 :- multifile('$lgt_logtalk.print_message_token'/3).